Cleaned up pdf to txt
