--- PDF_to_text/1.csv
@@ -1,15 +1,0 @@
-File Name,6-digit Numbers
-Bateman et al. - 2003 - Integrated predictive modelling simulations of bur.txt,99411 99411
-Garofalo et al. - 2006 - Access to sustained high-beta with internal transp.txt,122976
-Hannum et al. - 2001 - Comparison of high-mode predictive simulations usi.txt,77557 77559 81321 81329 81499 81507 82205 82788
-Holland - 2016 - Validation metrics for turbulent plasma transport.txt,128913 128913
-Kinsey et al. - 2002 - Simulations of internal transport barrier formatio.txt,92389 99696 95989
-Meneghini et al. - 2015 - Integrated modeling applications for tokamak experiments with OMFIT.txt,128913
-Meneghini et al. - 2017 - Self-consistent core-pedestal transport simulation.txt,145452 145452
-Onjun et al. - 2001 - Comparison of low confinement mode transport simul.txt,78281 78106 78283 78109 78283 78109 78283 78109 78281 78106 78283 78109
-Pankin et al. - 2013 - Validation of transport models using additive flux.txt,98889 98889 132016 132014 132017 132017 132017 98889
-Snyder et al. - 2002 - Edge localized modes and the pedestal A model bas.txt,97887
-Snyder et al. - 2012 - The EPED pedestal model and edge localized mode-su.txt,144977 144977 144977 145380 145380
-Snyder et al. - 2012 - The EPED pedestal model and edge localized mode-suppressed regimes Studies of quiescent H-mode and development(2).txt,144977 144977 144977 145380 145380
-Snyder et al. - Unknown - Development and validation of a predictive model for the pedestal height a ELMs Fusion.txt,132010 132010
-Van Zeeland et al. - 2012 - Alfvén eigenmode stability and fast ion loss in DI.txt,142111
